add nx cloud tree view item and update api calls
