+++ libs/shared/telemetry/src/lib/telemetry-types.ts
@@ -27,10 +27,10 @@ export type TelemetryEvents =
   | 'cloud.view-run'
   | 'cloud.explain-cipe-error'
   | 'cloud.fix-cipe-error'
-  | 'cloud.show-ai-task-fix-notification'
-  | 'cloud.show-ai-task-fix'
-  | 'cloud.apply-ai-task-fix'
-  | 'cloud.ignore-ai-task-fix'
+  | 'cloud.show-ai-fix-notification'
+  | 'cloud.show-ai-fix'
+  | 'cloud.apply-ai-fix'
+  | 'cloud.ignore-ai-fix'
   // graph
   | 'graph.show-all'
   | 'graph.show-affected'
